--- airbyte-integrations/bases/standard-source-test/src/main/java/io/airbyte/integrations/base/TestBasicTest.java
@@ -1,18 +1,0 @@
-package io.airbyte.integrations.base;
-
-import static org.junit.jupiter.api.Assertions.fail;
-
-import org.junit.jupiter.api.Test;
-
-public class TestBasicTest {
-
-  @Test
-  void myTest(){
-    System.out.println("true = " + true);
-  }
-
-  @Test
-  void myTest2(){
-    fail();
-  }
-}
