+++ airbyte-integrations/connectors/source-mysql/build.gradle
@@ -6,26 +6,24 @@ apply from: rootProject.file('tools/gradle/commons/integrations/image.gradle')
 apply from: rootProject.file('tools/gradle/commons/integrations/integration-test.gradle')
 
 dependencies {
-//    implementation project(':airbyte-config:models')
     implementation project(':airbyte-db')
     implementation project(':airbyte-integrations:bases:base-java')
     implementation project(':airbyte-protocol:models')
     implementation project(':airbyte-integrations:connectors:source-jdbc')
-//    implementation project(':airbyte-queue')
-    testImplementation project(':airbyte-test-utils')
 
     implementation 'mysql:mysql-connector-java:8.0.22'
-    testImplementation 'org.testcontainers:mysql:1.14.3'
+    // on rc version due to docker pull bug in current stable version.
+    // issue: https://github.com/airbytehq/airbyte/issues/493
+    testImplementation 'org.testcontainers:mysql:1.15.0-rc2'
 
     implementation 'org.apache.commons:commons-lang3:3.11'
 
-    // todo hook up to source lib
-//    integrationTestImplementation project(':airbyte-integrations:bases:destination-test-lib')
-//    integrationTestImplementation "org.testcontainers:postgresql:1.15.0-rc2"
+    integrationTestImplementation project(':airbyte-integrations:bases:standard-source-test')
+    testImplementation 'org.testcontainers:mysql:1.15.0-rc2'
 }
 
 application {
-    mainClass = 'io.airbyte.integrations.destination.postgres.JdbcSource'
+    mainClass = 'io.airbyte.integrations.source.mysql.MySqlSource'
 }
 
 buildImage.dependsOn(assemble)
