+++ airbyte-integrations/connectors/source-mysql/src/main/resources/spec.json
@@ -1,12 +1,27 @@
 {
-  "documentationUrl": "https://docs.airbyte.io/integrations/destinations/postgres",
+  "documentationUrl": "https://docs.airbyte.io/integrations/destinations/mysql",
   "connectionSpecification": {
     "$schema": "http://json-schema.org/draft-07/schema#",
-    "title": "Postgres Destination Spec",
+    "title": "MySql Source Spec",
     "type": "object",
-    "required": ["username", "jdbc_url"],
+    "required": ["host", "port", "database", "username"],
     "additionalProperties": false,
     "properties": {
+      "host": {
+        "description": "Hostname of the database.",
+        "type": "string"
+      },
+      "port": {
+        "description": "Port of the database.",
+        "type": "integer",
+        "minimum": 0,
+        "maximum": 65536,
+        "default": 3306,
+        "examples": ["3306"]
+      },
+      "database": {
+        "description": "Name of the database.",
+        "type": "string"
       },
       "username": {
         "description": "Username to use to access the database.",
@@ -14,10 +29,6 @@
       },
       "password": {
         "description": "Password associated with the username.",
-        "type": "string"
-      },
-      "jdbc_url": {
-        "description": "JDBC formatted url. See the standard <a href=\"https://docs.oracle.com/cd/E17952_01/connector-j-8.0-en/connector-j-reference-jdbc-url-format.html\">here</a>.",
         "type": "string"
       }
     }
