--- airbyte-integrations/connectors/source-mysql/src/test/resources/init_ascii.sql
@@ -1,8 +1,0 @@
-CREATE TABLE id_and_name(id INTEGER, name VARCHAR(200));
-
-INSERT INTO id_and_name
-    (id, name)
-VALUES
-    (1,'picard'),
-    (2, 'crusher'),
-    (3, 'vash');
